Move dev-server block logic to `webpack-common`
