+++ packages/dev-server2/package.json
@@ -16,7 +16,8 @@
   "bugs": "https://github.com/andywer/webpack-blocks/issues",
   "dependencies": {
     "webpack": "^2.1.0-beta.25",
-    "webpack-dev-server": "^2.1.0-beta.11"
+    "webpack-dev-server": "^2.1.0-beta.11",
+    "@webpack-blocks/webpack-common": "^0.3.0"
   },
   "devDependencies": {
     "standard": "^8.1.0"
