+++ packages/webpack-common/src/index.js
@@ -1,2 +1,3 @@
 export { default as defineConstants } from './defineConstants'
+export { default as devServer } from './devServer'
 export * from './webpack'
